+++ djpcms/views/docview.py
@@ -6,79 +6,54 @@ from django import http
 from django.shortcuts import render_to_response
 from django.template import loader, RequestContext
 from django.utils.safestring import mark_safe
+from django.conf.urls.defaults import url
 
 from unipath import FSPath as Path
 
 from djpcms.views.baseview import djpcmsview
-from djpcms.views.appview import ApplicationBase
+from djpcms.views.appsite import ApplicationBase
+from djpcms.views.appview import AppViewBase
+from djpcms.views.regex import RegExUrl
 
 
-
-class DocView(djpcmsview, ApplicationBase):
-    # Name the documents
-    name             = None
-    deflang          = 'en'
-    DOCS_PICKLE_ROOT = None
+class DocView(AppViewBase):
     editurl          = None
     
-    def get_root_code(self):
-        return self.name
+    def __init__(self, app, parent = None):
+        AppViewBase.__init__(self, parent = parent)
+        self.appmodel = app
+        
+class Index(DocView):
     
-    def make_urls(self):
-        '''
-        Create a tuple of urls
-        '''
-        from django.conf.urls.defaults import url
-        baseurl = self.baseurl
-        if baseurl == None:
-            return None
-        self.doc_index = '%s_doc_index' % self.name
-        b = baseurl[1:]
-        return (url(r'^%s$' % b, self.index),
-                url(r'^%s(?P<lang>[a-z-]+)/$' % b, self.language),
-                url(r'^%s(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/$' % b,
-                    self.document,
-                    {'url': ''},
-                    name = '%s_doc_index' % self.name),
-                url(r'^%s(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/(?P<url>[\w./-]*)/$' % b,
-                    self.document,
-                    name = '%s_doc_details' % self.name),
-                )
+    def __init__(self, app):
+        super(Index,self).__init__(app)
     
-    def get_path_args(self, lang, version):
-        return (lang, version, "_build", "json")
-    
-    def get_docroot(self, lang, version):
-        docroot = Path(self.DOCS_PICKLE_ROOT).child(*self.get_path_args(lang, version))
-        if not docroot.exists():
-            raise http.Http404()
-        return docroot 
-    
-    def index(self, request):
-        return http.HttpResponseRedirect('%s%s/dev/' % (self.baseurl,self.deflang))
-    
-    def language(self, request, lang):
-        return http.HttpResponseRedirect('%s%s/dev/' % (self.baseurl,lang))
+    def __call__(self, request, *args, **kwargs):
+        a = self.appmodel
+        return http.HttpResponseRedirect('%s%s/%s/' % (a.baseurl,a.deflang,a.defversion))
 
-    def get_template(self,page):
-        '''
-        given a page objects (which may be None)
-        return the template file for the get view
-        '''
-        return None
-            
-    def document(self, request, lang, version, url = None):
-        djp = self(request, lang = lang, version = version, url = url)
-        return djp.response()
+class Language(DocView):
     
-    def render(self, djp):
+    def __init__(self, app, parent):
+        super(Language,self).__init__(app,parent)
+    
+    def __call__(self, request, lang, *args, **kwargs):
+        a = self.appmodel
+        return http.HttpResponseRedirect('%s%s/%s/' % (a.baseurl,lang,a.defversion))
+
+class Document(DocView):
+    
+    def __init__(self, app, parent):
+        super(Document,self).__init__(app,parent)
+    
+    def handle_response(self, djp):
         request = djp.request
         args    = djp.urlargs
         page    = djp.page
         lang    = args.get('lang')
         version = args.get('version')
         url     = args.get('url')
-        docroot = self.get_docroot(lang, version)
+        docroot = self.appmodel.get_docroot(lang, version)
     
         # First look for <bits>/index.fpickle, then for <bits>.fpickle
         bits = url.strip('/').split('/') + ['index.fjson']
@@ -90,10 +65,11 @@ class DocView(djpcmsview, ApplicationBase):
                 raise http.Http404("'%s' does not exist" % doc)
 
         bits[-1] = bits[-1].replace('.fjson', '')
+        name  = self.appmodel.name
         namet = '-'.join([b for b in bits if b])
         template_names = [
-                          '%s/docs/%s.html' % (self.name,namet),
-                          '%s/docs/doc.html' % self.name,
+                          '%s/docs/%s.html' % (name,namet),
+                          '%s/docs/doc.html' % name,
                           'docs/%s.html' % namet, 
                           'docs/doc.html'
                           ]
@@ -113,6 +89,48 @@ class DocView(djpcmsview, ApplicationBase):
         
         return render_to_response(template_names,
                                   RequestContext(request,c))
+        
+
+
+class DocApplication(ApplicationBase):
+    deflang          = 'en'
+    defversion       = 'dev'
+    DOCS_PICKLE_ROOT = None
+    
+    def __init__(self, application_site, editavailable):
+        '''
+        Create a tuple of urls
+        '''
+        super(DocApplication,self).__init__(application_site, editavailable)       
+        self.root_application   = Index(self)
+        self.language           = Language(self,self.root_application)
+        self.document           = Document(self,self.language)
+        self.urls = (url(r'^$', self.root_application),
+                     url(r'^(?P<lang>[a-z-]+)/$', self.language),
+                     url(r'^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/$',
+                         self.document, {'url': ''},  name = '%s_index' % self.name),
+                     url(r'^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/(?P<url>[\w./-]*)/$',
+                         self.document, name = '%s_details' % self.name),
+                    )
+    
+    def get_root_code(self):
+        return self.name
+    
+    def get_path_args(self, lang, version):
+        return (lang, version, "_build", "json")
+    
+    def get_docroot(self, lang, version):
+        docroot = Path(self.DOCS_PICKLE_ROOT).child(*self.get_path_args(lang, version))
+        if not docroot.exists():
+            raise http.Http404()
+        return docroot 
+
+    def get_template(self,page):
+        '''
+        given a page objects (which may be None)
+        return the template file for the get view
+        '''
+        return None
         
     def bodybits(self):
         if self.editurl: